add permision

+++ LuxeAssist/request/views.py
@@ -8,8 +8,8 @@ from .models import RequestComment
 
 
 def add_Request_view(request:HttpRequest, service_id):
-        #if not request.user.is_authenticated:
-              # return render(request, "", status=401)
+    if not request.user.is_authenticated:
+        return render(request, "main/user_not_found.html", status=401)
         
     service = Service.objects.get(id = service_id)  
     if request.method == "POST":
@@ -19,8 +19,8 @@ def add_Request_view(request:HttpRequest, service_id):
         
 
 def user_requests_view(request: HttpRequest):
-    # if request.user.is_authenticated and not request.user.is_superuser and request.user.is_staff :
-    #    return render(request,"main/user_not_found.html", status=401)
+    if not request.user.is_authenticated :
+       return render(request,"main/user_not_found.html", status=401)
    # try:
         
     requests = Request.objects.filter(user= request.user)
@@ -32,8 +32,8 @@ def user_requests_view(request: HttpRequest):
     
 
 def concierge_requests_view(request: HttpRequest): 
-    # if  not request.user.is_authenticated and not request.user.groups.exists :
-    #     return render(request,"main/user_not_found.html", status=401)
+    if  not request.user.is_authenticated and not request.user.groups.exists :
+        return render(request,"main/user_not_found.html", status=401)
    # try:
     services=Service.objects.filter(user=request.user)
     requests = Request.objects.filter(service__in= services)
@@ -45,13 +45,13 @@ def concierge_requests_view(request: HttpRequest):
     
 
 def admin_requests_view(request: HttpRequest):
-    # if request.user.is_superuser and request.user.is_staff :
-    #         return render(request,"main/user_not_found.html", status=401)
+    if not request.user.is_superuser :
+       return render(request,"main/user_not_found.html", status=401)
   # try: 
-      requests = Request.objects.order_by('-date')
+    requests = Request.objects.order_by('-date')
      
      
-      return render(request ,"request/admin_requests_view.html" , {"requests" : requests})
+    return render(request ,"request/admin_requests_view.html" , {"requests" : requests})
    #except Exception as e:
 
        # return render(request, "")
@@ -75,8 +75,8 @@ def request_details_view(request: HttpRequest,requsets_id):
      
 
 def cancel_request_view(request: HttpRequest, requset_id):
-    # if request.user.is_superuser and request.user.is_staff :
-    #         return render(request,"main/user_not_found.html", status=401)
+    if request.user.is_superuser and request.user.is_staff :
+       return render(request,"main/user_not_found.html", status=401)
     
     try:
         requests = Request.objects.get(id = requset_id)
@@ -91,8 +91,8 @@ def update_price_view(request:HttpRequest, requests_id):
 
     requests = Request.objects.get(id = requests_id) 
 
-    if  not request.user.is_superuser and request.user.is_superuser and request.user.is_staff :
-            return render(request,"main/user_not_found.html", status=401)
+    if request.user.is_superuser and request.user.is_staff :
+        return render(request,"main/user_not_found.html", status=401)
      
     if request.method == "POST":
         requests.request_price = request.POST["request_price"]
@@ -101,8 +101,8 @@ def update_price_view(request:HttpRequest, requests_id):
 
 
 def add_status_view(request: HttpRequest, requests_id):
-    # if request.user.is_authenticated and not request.user.groups.exists :
-    #     return render(request,"main/user_not_found.html", status=401)
+    if not request.user.is_authenticated and not request.user.groups.exists :
+       return render(request,"main/user_not_found.html", status=401)
 
     requests = Request.objects.get(id = requests_id)
 
@@ -119,8 +119,8 @@ def new_requestConcierge_view(request: HttpRequest , requset_id):
 
 
 def delete_request_admin_view(request: HttpRequest, requset_id):
-    # if not request.user.is_superuser and request.user.is_staff :
-    #     return render(request,"main/user_not_found.html", status=401)
+    if not request.user.is_superuser :
+        return render(request,"main/user_not_found.html", status=401)
 
     requsets=Request.objects.get(id=requset_id)
     requsets.delete()
